% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{ate}
\alias{ate}
\title{Estimate non-overlap bounds for the Average Treatment Effect (ATE)}
\usage{
ate(
  data,
  X,
  A,
  Y,
  learners_trt = c("SL.glm"),
  learners_outcome = c("SL.glm"),
  thresholds = c(10^seq(-4, -1, 0.05)),
  smoothness = 0.01,
  alpha = 0.05,
  outer_folds = 5,
  inner_folds = 5,
  bootstrap = TRUE,
  bootstrap_draws = 1000,
  nuisance = NULL
)
}
\arguments{
\item{data}{data frame containing data estimating ATE bounds}

\item{X}{vector of covariate column names}

\item{A}{name of column containing binary treatment indicator}

\item{Y}{name of column containing binary outcome variable}

\item{learners_trt}{SuperLearner learners to use to estimate propensity score model}

\item{learners_outcome}{SuperLearner learners to use to estimate outcome model}

\item{thresholds}{vector of propensity score thresholds}

\item{smoothness}{tuning parameter controlling smoothness of the indicator function approximations (smaller values imply a less smooth approximation)}

\item{alpha}{significance level of pointwise and uniform confidence intervals (default 5\%)}

\item{outer_folds}{Number of folds in outer cross-fitting loop}

\item{inner_folds}{Number of folds used by SuperLearner within each outer cross-fitting loop}

\item{bootstrap}{whether to use multiplier bootstrap to compute uniform confidence sets}

\item{bootstrap_draws}{number of multiplier bootstrap draws}

\item{nuisance}{list containing estimated nuisance parameters (optional)}
}
\value{
object with class "atebounds"
}
\description{
Estimate non-overlap bounds for the Average Treatment Effect (ATE)
}
\examples{
dat <- simulate_ate_example(
  seed = 1,
  N = 5e2,
  alpha = 3,
  beta = 0.1,
  gamma = 1
)

bounds <- ate(
  dat,
  X = c("X1", "X2"), A = "A", Y = "Y",
  thresholds = c(10^seq(-3, -0.5, 0.1)),
  smoothness = c(0.005)
)

}
