% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdrf.R
\name{cdrf_bounds}
\alias{cdrf_bounds}
\title{Estimate non-overlap bounds for the Causal Dose-Response Function (CDRF)}
\usage{
cdrf_bounds(
  data,
  X,
  A,
  Y,
  learners_trt = c("SL.glm"),
  learners_outcome = c("SL.glm"),
  trt_grid = seq(0, 1, 0.1),
  thresholds = c(10^seq(-4, -1, 0.05)),
  smoothness = 0.01,
  bw = 0.1,
  alpha = 0.05,
  outer_folds = 5,
  inner_folds = 5,
  bootstrap = TRUE,
  bootstrap_draws = 1000,
  nuisance = NULL
)
}
\arguments{
\item{data}{data frame containing data estimating CDRF bounds}

\item{X}{vector of covariate column names}

\item{A}{name of column containing binary treatment indicator}

\item{Y}{name of column containing outcome variable (bounded between zero and one)}

\item{learners_trt}{SuperLearner learners for estimating propensity
score model.}

\item{learners_outcome}{SuperLearner learners for estimtaing outcome
model.}

\item{trt_grid}{grid of treatment values}

\item{thresholds}{vector of propensity score thresholds}

\item{smoothness}{tuning parameter controlling smoothness of the indicator
unction approximations (smaller values imply a less smooth approximation)}

\item{alpha}{significance level of pointwise and
uniform confidence intervals (default 5\%)}

\item{outer_folds}{Number of folds in outer cross-fitting loop}

\item{inner_folds}{Number of folds used by SuperLearner within
each outer cross-fitting loop}

\item{bootstrap}{whether to use multiplier bootstrap to compute
uniform confidence sets}

\item{bootstrap_draws}{number of multiplier bootstrap draws}

\item{nuisance}{list containing estimated nuisance parameters (optional)}

\item{bandwidth}{kernel bandwidth}
}
\value{
A list of class \code{cdrfbounds} containing the following elements:
\describe{
\item{bounds}{List containing estimated bounds for each smoothness parameter.}
\item{smoothness}{Vector of smoothness parameters.}
\item{thresholds}{Vector of propensity score thresholds.}
\item{onestep}{One-step point estimate and confidence interval for CDRF.}
\item{alpha}{Significance level.}
\item{N}{Number of observations.}
\item{K}{Number of propensity score thresholds.}
\item{nuisance}{propensity score and conditional mean outcome predictions.}
}
}
\description{
Estimate non-overlap bounds for the Causal Dose-Response Function (CDRF)
}
\details{
This function estimates non-overlap bounds for the Causal Dose-Response
Function (CDRF) on a user-specified grid of exposure values, propensity score
thresholds and smoothness, and tuning parameter values.

By default, 95\% uniform confidence sets are formed. You can configure the
significance level with the \code{alpha} argument; the uniform confidence
sets are designed to be valid with probability \eqn{(1 - \alpha) \times 100\%}.

An important tuning parameter is the \code{smoothness} argument, which
controls the smoothness of an inner approximation of certain indicator
functions that arise in the definition of the non-overlap bounds. In practice,
a small value (like \code{10e-2}) can typically be used. We
recommend trying several small values, like \code{10e-3}, \code{10e-2},
and \code{10e-1} in a sensitivity analysis.
}
\examples{
dat <- simulate_cdrf_example(
  seed = 1,
  N = 5e2,
  alpha = 3,
  beta = 0.1,
  gamma = 1
)

bounds <- cdrf_bounds(
  dat,
  X = c("X1", "X2"), A = "A", Y = "Y",
  thresholds = c(10^seq(-3, -0.5, 0.1)),
  smoothness = c(0.005)
)

}
\seealso{
\link{summary.cdrfbounds}

\link{plot.cdrfbounds}
}
