% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ci}
\alias{ci}
\title{Compute CI from influence function estimate}
\usage{
ci(estimate, eif, N, alpha = 0.05)
}
\arguments{
\item{estimate}{point estimtae}

\item{eif}{estimate of EIF (vector)}

\item{N}{sample size}

\item{alpha}{significance level}
}
\description{
Compute CI from influence function estimate
}
