% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eifs.R
\name{eif_ate}
\alias{eif_ate}
\title{Efficient Influence Function (EIF) for the Average Treatment Effect (ATE) parameter}
\usage{
eif_ate(A, Y, mu0, mu1, pi)
}
\arguments{
\item{A}{binary treatment indicator}

\item{Y}{binary outcome}

\item{mu0}{conditional mean of Y given A = 0}

\item{mu1}{conditional mean of Y given A = 0}

\item{pi}{propensity score P(A = 1 | X)}
}
\value{
vector of efficient influence function values
}
\description{
Efficient Influence Function (EIF) for the Average Treatment Effect (ATE) parameter
}
