% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eifs.R
\name{eif_upper}
\alias{eif_upper}
\title{Efficient Influence Function (EIF) for the smooth lower ATE bound}
\usage{
eif_upper(A, Y, mu0, mu1, pi, threshold, smoothness)
}
\arguments{
\item{A}{binary treatment indicator}

\item{Y}{binary outcome}

\item{mu0}{conditional mean of Y given A = 0}

\item{mu1}{conditional mean of Y given A = 0}

\item{pi}{propensity score P(A = 1 | X)}

\item{threshold}{propensity score threshold}

\item{smoothness}{smooth approximation tuning parameter}
}
\value{
vector of efficient influence function values
}
\description{
Efficient Influence Function (EIF) for the smooth lower ATE bound
}
