% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{multiplier_bootstrap}
\alias{multiplier_bootstrap}
\title{Apply multiplier bootstrap for lower and upper effect bounds}
\usage{
multiplier_bootstrap(
  lower,
  upper,
  eif_lower,
  eif_upper,
  draws = 1000,
  alpha = 0.05
)
}
\arguments{
\item{lower}{point estimates of lower bound for set of propensity score thresholds (vector)}

\item{upper}{point estimates of upper bound for set of propensity score thresholds (vector)}

\item{eif_lower}{efficient influence function estimates for lower bound (matrix)}

\item{eif_upper}{efficient influence function estimates for upper bound (matrix)}

\item{draws}{number of bootstrap values}

\item{alpha}{significance level, leading to (1 - alpha)\% uniform confidence set}
}
\description{
Apply multiplier bootstrap for lower and upper effect bounds
}
