% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{onestep}
\alias{onestep}
\title{Compute one-step estimator for one-dimensional target parameter}
\usage{
onestep(plugin, eif)
}
\arguments{
\item{plugin}{plugin point estimate}

\item{eif}{EIF vector}
}
\description{
Compute one-step estimator for one-dimensional target parameter
}
