% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cdrfbounds}
\alias{plot.cdrfbounds}
\title{Plot estimated CDRF non-overlap bounds}
\usage{
\method{plot}{cdrfbounds}(
  x,
  smoothness = x$smoothness[1],
  point_estimate = FALSE,
  ylim = NA,
  xlab = "Propensity Score Threshold",
  ylab = "ATE",
  legend_position = "bottomright",
  bounds_color = "black",
  point_estimate_color = "blue",
  ...
)
}
\arguments{
\item{x}{object of type "cdrfbounds"}

\item{smoothness}{which smoothness tuning parameter bounds to plot; set to NA to print all}

\item{point_estimate}{whether to plot point estimate and 95\% confidence interval}

\item{ylim}{limits of y axis}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{legend_position}{where to show legend (set to "none" to hide legend)}

\item{bounds_color}{color of point estimate and 95\% confidence interval}

\item{point_estimate_color}{color of point estimate and 95\% confidence interval}

\item{...}{additional arguments passed to plot.default}
}
\description{
Plot estimated CDRF non-overlap bounds
}
