% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.atebounds}
\alias{print.atebounds}
\title{Print estimated ATE non-overlap bounds}
\usage{
\method{print}{atebounds}(x, ...)
}
\arguments{
\item{x}{object of class "atebounds"}

\item{...}{further arguments (not used)}
}
\description{
Print estimated ATE non-overlap bounds
}
