% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.atebounds}
\alias{print.summary.atebounds}
\title{Print ATE non-overlap bounds summary}
\usage{
\method{print}{summary.atebounds}(x, ...)
}
\arguments{
\item{x}{object of class "summary.atebounds"}

\item{...}{additional arguments}
}
\description{
Print ATE non-overlap bounds summary
}
