% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{s_gt}
\alias{s_gt}
\title{Smooth approximation of the indicator function I\link{x > t} with smoothness gamma
Note that, by construction, s(x) <= I\link{x > t}.}
\usage{
s_gt(x, t, gamma)
}
\description{
Smooth approximation of the indicator function I\link{x > t} with smoothness gamma
Note that, by construction, s(x) <= I\link{x > t}.
}
