% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{s_lt_dot}
\alias{s_lt_dot}
\title{Derivative of smooth approximation of indicator function I\link{X > t} with smoothness gamma}
\usage{
s_lt_dot(x, t, gamma)
}
\description{
Derivative of smooth approximation of indicator function I\link{X > t} with smoothness gamma
}
