% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_ate_example}
\alias{simulate_ate_example}
\title{Simulate from a simple illustrative data-generating process designed
to illustrate ATE non-overlap bounds}
\usage{
simulate_ate_example(
  seed = 1,
  N = 1000,
  alpha = 1,
  beta = 0.05,
  gamma = 1,
  binary_outcome = TRUE
)
}
\arguments{
\item{seed}{random number seed}

\item{N}{number of observations}

\item{alpha}{strength of non-overlap violations (higher alpha implies more extreme non-overlap violation)}

\item{beta}{proportion of population subject to non-overlap violation}

\item{gamma}{controls magnitude of the effect}

\item{binary_outcome}{specifies whether to generate a binary (default) or continuous outcome}
}
\value{
data frame with columns X1, X2, A, and Y
}
\description{
Simulate from a simple illustrative data-generating process designed
to illustrate ATE non-overlap bounds
}
