% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{tmle_smooth_ate}
\alias{tmle_smooth_ate}
\title{TMLE algorithm for non-overlap ATE bound parameters}
\usage{
tmle_smooth_ate(
  A,
  Y,
  mu0,
  mu1,
  pi,
  threshold,
  smoothness,
  parameter = "trimmed",
  maxiter = 25,
  verbose = FALSE
)
}
\description{
TMLE algorithm for non-overlap ATE bound parameters
}
