% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.atebounds}
\alias{summary.atebounds}
\title{Summarize ATE non-overlap bounds}
\usage{
\method{summary}{atebounds}(object, ...)
}
\arguments{
\item{object}{object of type "atebounds"}

\item{...}{additional arguments}
}
\description{
Summarize ATE non-overlap bounds
}
